requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.2.0